+++ packages/core/src/elements/Button/FeButton.tsx
@@ -43,7 +43,7 @@ export const FeButton = forwardRef<HTMLButtonElement, ButtonProps>((props, ref) 
       {...restProps}
     >
       {children}
-      {loading && <FeLoader />}
+      {loading && <FeLoader size={size === 'small' ? 18 : 24} />}
     </button>
   );
 });