+++ packages/demo-saas/src/notifications-example/index.tsx
@@ -1,0 +1,10 @@
+import React, { FC } from 'react';
+import { Notifications } from '@frontegg/react-notifications';
+
+export const NotificationsExample: FC = () => {
+  return (
+    <div style={{ textAlign: 'center' }}>
+      <Notifications />
+    </div>
+  );
+};