+++ packages/demo-saas/src/withFrontegg.tsx
@@ -1,10 +1,9 @@
 import React, { ComponentType } from 'react';
-import { ContextOptions, fronteggElements, FronteggProvider, PluginConfig } from '@frontegg/react-core';
+import { ContextOptions, FronteggProvider, PluginConfig } from '@frontegg/react-core';
 import { AuthPlugin } from '@frontegg/react-auth';
 import { IntegrationsPlugin } from '@frontegg/react-integrations';
-
-import { uiLibrary as SemanticLibrary } from '@frontegg/react-elements-semantic';
-import { uiLibrary as MaterialLibrary } from '@frontegg/react-elements-material-ui';
+import { NotificationsPlugin } from '@frontegg/react-notifications';
+import { uiLibrary } from '@frontegg/react-elements-material-ui';
 
 const developmentHosts = ['localhost', 'local.frontegg.com'];
 const host =
@@ -17,34 +16,10 @@ const contextOptions: ContextOptions = {
   requestCredentials: 'include',
 };
 
-const plugins: PluginConfig[] = [AuthPlugin(), IntegrationsPlugin()];
+const plugins: PluginConfig[] = [AuthPlugin({}), IntegrationsPlugin(), NotificationsPlugin({})];
 
 export const withFrontegg = (Component: ComponentType<any>) => () => (
-  <FronteggProvider
-    debugMode
-    context={contextOptions}
-    plugins={plugins}
-    uiLibrary={
-      {
-        // ...MaterialLibrary,
-        // Input: SemanticLibrary.Input,
-        // Dialog: SemanticLibrary.Dialog,
-        // Form: SemanticLibrary.Form,
-        // Select: SemanticLibrary.Select,
-        // ...(localStorage.getItem('library') === 'material' ? MaterialLibrary : SemanticLibrary),
-        // Tag: fronteggElements.Tag
-        // Input: SemanticLibrary.Input,
-        // Form: SemanticLibrary.Form,
-        // Dialog: MaterialLibrary.Dialog,
-        // Loader: fronteggElements.Loader,
-        // Button: fronteggElements.Button,
-        // ...MaterialLibrary,
-        // ...SemanticLibrary,
-        // SwitchToggle: SemanticLibrary.SwitchToggle
-        // Tabs: SemanticLibrary.Tabs
-      }
-    }
-  >
+  <FronteggProvider debugMode context={contextOptions} plugins={plugins} uiLibrary={uiLibrary}>
     <Component />
   </FronteggProvider>
 );