+++ packages/rest-api/src/constants.ts
@@ -11,3 +11,4 @@ export const EVENT_RESOURCES_URL_V1 = '/event/resources/configurations/v1';
 export const TEAMS_USERS_SERVICE_URL_V1 = '/team/resources/members/v1';
 export const TEAMS_ROLES_SERVICE_URL_V1 = '/team/resources/roles/v1';
 export const TEAMS_STATS_SERVICE_URL_V1 = '/team/resources/stats/v1/members';
+export const NOTIFICATIONS_SERVICE_URL = '/notification';