+++ packages/rest-api/src/index.ts
@@ -2,6 +2,7 @@ import * as auth from './auth';
 import * as teams from './teams';
 import * as metadata from './metadata';
 import * as reports from './reports';
+import * as notifications from './notifications';
 import * as fetch from './fetch';
 import * as integrations from './integrations';
 
@@ -13,6 +14,7 @@ export * from './teams/interfaces';
 export * from './metadata/interfaces';
 export * from './reports/interfaces';
 export * from './integrations/interfaces';
+export * from './notifications/interfaces';
 
 const api = {
   auth,
@@ -20,6 +22,7 @@ const api = {
   metadata,
   reports,
   integrations,
+  notifications,
 };
 
 export { fetch, ContextHolder, FronteggContext, api };